소지금액 int로 변경, 지역 구매 구현
1. 소지금액 int로 변경
- txt파일 데이터를 float로 읽어오면 오차 포함되어 3을 읽어도 2.999 이런 식으로 들어와서 해당 오차를 없애기 위해 소지 금액을 int로 사용 중

2. 지역 구매 구현
- 지역을 구매하면, 구매 싱크를 서버에서 보내고, 그 정보를 통해 클라에서 구매 유무를 확인하고, 구매 시 해당 지역 이름 색을 변경함으로 구분했음

-> 지금은 무조건 구매만 할 수 있도록 구현했기 때문에, 통행료, 건물 건축 등 상황은 따로 구현해야 함

+++ BlueMarble_Client/GameManager.cpp
@@ -48,7 +48,7 @@ vector<int>* GameManager::GetUserPositionVector()
 	return &userPositionVector;
 }
 
-vector<float>* GameManager::GetUserMoneyVector()
+vector<int>* GameManager::GetUserMoneyVector()
 {
 	return &userMoneyVector;
 }
